app/requirements.txt
jsonlines==3.1.0
scikit-learn==1.0.2
WTForms 
flask_wtf==1.0.1
spacy==3.4.1
spacy-alignments==0.8.6
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-transformers==1.1.8
pyparsing==3.0.9
PyPDF2==2.11.1
nlp==0.4.0
nltk==3.7
numpy==1.23.4
pandas==1.5.0
sympy==1.11.1